requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Pygments==1.6
Werkzeug==0.9.4
argparse==1.1
bpython==0.12
httplib2==0.8
itsdangerous==0.23
oauthlib==0.6.0
peewee==2.1.4
python-instagram==0.8.0
requests==1.2.3
requests-oauth==0.4.1
requests-oauthlib==0.4.0
simplejson==3.3.1
wsgiref==0.1.2
